Type organization and cleanup
